+++ Monofoxe.Playground/Monofoxe.Playground/ECDemo/ECDemoFactory.cs
@@ -29,7 +29,14 @@ namespace Monofoxe.Playground.ECDemo
 				position.Position = new Vector2(Random.Next(100, 700), Random.Next(100, 500));
 			}
 
-			new Player(layer, new Vector2(400, 300));
+			var player = new Player(layer, new Vector2(400, 300));
+
+			// Now player will be drawn below all the bots, even though he was created last.
+			// Reaordering puts him on the top of entity list.
+			layer.ReorderEntityToTop(player);
+			//layer.ReorderEntityToBottom(player); // Will have no effect.
+			//layer.ReorderEntity(player, 2); // Player will be updated and drawn only after two entities above him.
+
 
 			new ECDemo(layer);